Final refactoring - Deliver price range
