+++ src/fr/vergne/qualitemeritis/OfferBestSeatsInPriceRangeTest.java
@@ -376,26 +376,6 @@ class OfferBestSeatsInPriceRangeTest {
 		}
 	}
 
-	static class PriceRange {
-
-		private final Price priceMin;
-		private final Price priceMax;
-
-		public PriceRange(Price priceMin, Price priceMax) {
-			this.priceMin = priceMin;
-			this.priceMax = priceMax;
-		}
-
-		public boolean includes(Price price) {
-			if (this.priceMin.currency() == price.currency()) {
-				return this.priceMin.value() <= price.value()//
-						&& this.priceMax.value() >= price.value();
-			} else {
-				return false;
-			}
-		}
-	}
-
 	static class SuggestionSystem {
 
 		private final Collection<Seat> seats;
